0.2.3 - Configured MFA
Added routes for verifyOTP. Installed packages for nodemailer services.
Created an OTP model.
Added in middleware to assemble and send out mail from an an address with injected env credentials.
Configured authController with a sendMail method that uses the middleware. Additionally, added in the export for OTP verification.

Notes:
OTPs are hashed so it is secure in the database.
OTP's are deleted on use and no more than one per user can exist.
sendMail is called from the Login exported route when a valid user signs in.